Split user auth and user action interfaces
