+++ backend/app/api/v1/auth/__init__.py
@@ -1,8 +1,8 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 from fastapi import APIRouter
-from backend.app.api.v1.auth.user import router as user_router
+from backend.app.api.v1.auth.auth import router as auth_router
 
-router = APIRouter(prefix='/auth', tags=['用户管理'])
+router = APIRouter(prefix='/auth', tags=['认证'])
 
-router.include_router(user_router, prefix='/users')
+router.include_router(auth_router, prefix='/users')
